\name{FrenchTriageR-package}
\alias{FrenchTriageR}
\title{Implémentation de l'Algorithme de Triage Médical FRENCH}
\description{
Package complet pour le triage automatisé selon le protocole FRENCH.
}
\details{
Fonctionnalités clés :
\itemize{
  \item Import des données depuis Excel
  \item Prétraitement automatisé
  \item Classification selon 6 niveaux d'urgence
  \item Génération de rapports
}
}
\author{
Edouard Lansiaux <edouard.lansiaux@orange.fr>
}
